+++ RELEASE_NOTES.md
@@ -1,4 +1,8 @@
-#### 2.4.4 - Unreleased
+#### 2.4.5 - February 19 2018
+* Add an optional parameter for the `System.Text.Encoding` to use when reading data to the CSV, HTML, and Json providers. This parameter is called `encoding` and should be present on all Load and AsyncLoad methods.
+* Fix handling of multipart form data payloads whose size exceeded ~80k bytes.
+
+#### 2.4.4 - January 20 2018
 * Fix parsing of unquoted HTML attributes containing URLs.
 * Fixed HTTP form body url encoding.